maint(polys): add sympy.polys.matrices package
Refactor domainmatrix.py into a package sympy.polys.matrices
