+++ sympy/polys/matrices/__init__.py
@@ -1,0 +1,5 @@
+from .domainmatrix import DomainMatrix
+
+__all__ = [
+    'DomainMatrix',
+]
